#1579 Adapt the sbt bridge for the new error message
